+++ backend/packages/workspaces/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/workspaces
 
+## 0.27.12
+
+### Patch Changes
+
+- Updated dependencies [b25732f]
+  - @cat-factory/contracts@0.293.0
+  - @cat-factory/kernel@0.286.0
+
 ## 0.27.11
 
 ### Patch Changes